[WIP] whisper refacto to support language output.
